Fix package declaration mapping & add Java Version option
Files were being moved to the correct directory but the package
declarations at the top weren't changing. This will assume all package
declarations in a file should be changed to the new package we're
mapping to.

I also added a Java Version option to the Extractor because the Spigot
codebase uses Java 8. It defaults to Java 6.

Signed-off-by: Jadon Fowler <j@jadon.io>